course/samples/hello_world/semantic_kernel/requirements.txt
semantic-kernel
httpx
python-dotenv
